split the retrieval of data and aco method
