+++ aco_method.py
@@ -5,41 +5,11 @@
 import sys
 import random
 import math
-import ccxt
 import pandas as pd
-import ta
-from ta.trend import SMAIndicator, EMAIndicator
 
-def get_data(exchange_name, symbol, window, sma_window_size, min_ema_window_size, max_ema_window_size, ema_increment):
-    '''
-    Extract daily cryptocurrency stock data
-
-    @param exchange_name: Exchange market to retrieve data
-    @param symbol: Cryptocurrency symbol to retrieve data
-    @param window: Period for data retrieval
-    @param sma_window_size: Window size of SMA for data retrieval
-    @param min_ema_window_size: Minimum window size of EMA for data retrieval
-    @param max_ema_window_size: Maximum window size of EMA for data retrieval
-    @param ema_increment: Window size increments for EMA
-    '''
-    # Retrieve data from the exchange market for a specific cryptocurrency symbol
-    exchange = getattr(ccxt, exchange_name)()
-    ohlcv = exchange.fetch_ohlcv(symbol, '1d', limit=window, since=1620691200000)
-
-    # Create a table for the data
-    df = pd.DataFrame(ohlcv, columns=['timestamp', 'open', 'high', 'low', 'close', 'volume'])
-    df['timestamp'] = pd.to_datetime(df['timestamp'], unit='ms')
-    
-    # Add SMA data into the table
-    sma_indicator = SMAIndicator(df['close'], sma_window_size, True)
-    df[f'sma-{sma_window_size}'] = sma_indicator.sma_indicator()
-
-    # Add EMA data into the table
-    for i in range(min_ema_window_size, max_ema_window_size + 1, ema_increment):
-        ema_indicator = EMAIndicator(df['close'], i, True)
-        df[f'ema-{i}'] = ema_indicator.ema_indicator()
-    
-    # Return the data table
+def csv_to_df(filename):
+    df = pd.read_csv(filename)
+    print(df)
     return df
 
 def buy(wallet, price, amount):
@@ -68,27 +38,30 @@ def sell(wallet, price, amount):
     new_trigger = "SELL"
     return new_wallet, new_amount, new_trigger
 
-def run_simulation(num_ants, Q, period, fix_sma, min_ema, max_ema, ema_increment):
+def run_simulation(num_ants, Q):
     '''
     Set a sell trigger and sell cryptocurrency
 
     @param num_ants: Number of ants
     @param Q: Pheremone constant
     @param period: Period of data
-    @param fix_sma: Window size of SMA 
-    @param min_ema: Minimum window size of EMA
-    @param max_ema: Maximum window size of EMA
-    @param ema_increment: Window size increments for EMA
     '''
     # Random seed to provide same results each run
     random.seed(100)
 
     # Retrive data from exchange market
-    data = get_data('kraken', 'BTC/AUD', int(period), int(fix_sma), int(min_ema), int(max_ema), int(ema_increment))
+    data = csv_to_df('aco_data.csv')
     
     # Initialise variables
     best_ema, best_wallet = 0, 0
     initial_wallet = 100
+
+    # Fixed based on aco_data.csv
+    period = 720
+    fix_sma = 50
+    min_ema = 5
+    max_ema = 100
+    ema_increment = 5
     
     # Create probabilistic array of EMA periods
     ema_periods = []
@@ -181,30 +154,19 @@ def run_simulation(num_ants, Q, period, fix_sma, min_ema, max_ema, ema_increment
     print(results.to_string(index=False))
     print(f'Best overall: EMA-{best_ema}, ${best_wallet}')
     #results.to_csv(f'aco_results.csv', index=False)
-    #data.to_csv(f'aco_data.csv', index=False)
 
 if __name__ == "__main__":
-    if len(sys.argv) != 8:
-        print(f'Usage: python3 {__file__} num_ants pheremone_constant period fixed_sma_window min_ema_window max_ema_window ema_increment')
+    if len(sys.argv) != 3:
+        print(f'Usage: python3 {__file__} num_ants pheremone_constant')
         print('num_ants: Number of ants to generate\n'
-              'pheremone_constant: Probability constant for updating ACO (value from 0 to 1)\n'
-              'period: Period of dataset\n'
-              'fixed_sma_window: SMA window size that is fixed\n'
-              'min_ema_window: Minimum EMA window size\n'
-              'max_ema_window: Maximum EMA window size\n'
-              'ema_increment: Step increment for EMA window size')
+              'pheremone_constant: Probability constant for updating ACO (value from 0 to 1)\n')
         sys.exit(1)
 
     num_ants = int(sys.argv[1])
     pheremone_constant = float(sys.argv[2])
-    period = int(sys.argv[3])
-    fixed_sma_window = int(sys.argv[4])
-    min_ema_window = int(sys.argv[5])
-    max_ema_window = int(sys.argv[6])
-    ema_increment = int(sys.argv[7])
 
     if pheremone_constant > 1:
         print(f'Pheremone value of {sys.argv[2]} is not a valid input. Please enter a value from 0 to 1.')
         sys.exit(1)
 
-    run_simulation(num_ants, pheremone_constant, period, fixed_sma_window, min_ema_window, max_ema_window, ema_increment)
+    run_simulation(num_ants, pheremone_constant)
